% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_in_zip.R
\name{read_excel_from_zip}
\alias{read_excel_from_zip}
\title{Read The Excel Files In A Zip Archive}
\usage{
read_excel_from_zip(zip_file)
}
\arguments{
\item{zip_file}{path to a zip file that contains xls & xlsx files to be read}
}
\value{
a list of data.frame
}
\description{
Read The Excel Files In A Zip Archive
}
\examples{
read_excel_from_zip(system.file(package="zipofxlsxreader","datasets.zip"))
}
